+++ src/lazyPages.js
@@ -41,3 +41,5 @@ export const VrijwilligersSjabloonForm = lazy(() => import('@/pages/Vrijwilliger
 export const VrijwilligersDataQuality = lazy(() => import('@/pages/Vrijwilligers/VrijwilligersDataQuality'));
 export const VrijwilligersRelationshipQuality = lazy(() => import('@/pages/Vrijwilligers/VrijwilligersRelationshipQuality'));
 export const Vrijwillig = lazy(() => import('@/pages/Vrijwillig/Vrijwillig'));
+export const TaakuitlegList = lazy(() => import('@/pages/Taakuitleg/TaakuitlegList'));
+export const TaakuitlegForm = lazy(() => import('@/pages/Taakuitleg/TaakuitlegForm'));